+++ srcs/parse/token_utils.c
@@ -6,20 +6,12 @@
 /*   By: jiskim <jiskim@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/14 22:00:14 by jiskim            #+#    #+#             */
-/*   Updated: 2022/03/31 01:46:08 by jiskim           ###   ########.fr       */
+/*   Updated: 2022/04/04 03:13:54 by jiskim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-/**
- * @brief
- *
- * @param data nullable
- * @return t_token*
- *
- * << >> < > | 그 외
- */
 t_token	*new_token(char *data)
 {
 	t_token	*new;
@@ -43,13 +35,9 @@ t_token	*new_token(char *data)
 	return (new);
 }
 
-/**
- * 수정된 코드는 모든 data를 그대로 사용하지 않으므로 무조건 free 해줍니다.
- * WORD type도 ast에 들어갈 때 realloc으로 새로운 문자열이 만들어집니다.
- */
 void	free_token_list(t_token *list)
 {
-	t_token *tmp;
+	t_token	*tmp;
 
 	tmp = list;
 	while (list)
